+++ packages/liferay-theme-tasks/index.js
@@ -1,16 +1,30 @@
 'use strict';
 
 var _ = require('lodash');
+var async = require('async');
 var doctor = require('./lib/doctor');
 var glob = require('glob');
+var lfrThemeConfig = require('./lib/liferay_theme_config');
 var path = require('path');
 var plugins = require('gulp-load-plugins')();
+var RegisterHooks = require('./lib/register_hooks');
 var versionControl = require('./lib/version_control.js');
+
+var gutil = plugins.util;
+
+var themeConfig = lfrThemeConfig.getConfig();
 
 module.exports.registerTasks = function(options) {
 	options = require('./lib/options')(options);
 
 	var gulp = options.gulp;
+
+	gulp.tasks = {};
+
+	RegisterHooks.hook(gulp, {
+		hookFn: options.hookFn,
+		hookModules: themeConfig ? themeConfig.hookModules : null
+	});
 
 	gulp = options.gulp = plugins.help(gulp);